LayoutTests/imported/w3c/web-platform-tests/custom-elements/reactions/Animation-expected.txt


PASS Animation.animate must enqueue an attributeChanged reaction when it adds the observed style attribute
PASS Animation.animate must enqueue an attributeChanged reaction when it mutates the observed style attribute
PASS Animation.animate must not enqueue an attributeChanged reaction when it mutates the style attribute but the style attribute is not observed
